Split e2e test files by role and rename recorder
The two-file e2e suite had grown unwieldy: harness_test.go mixed six
unrelated concerns. Reorganize so the directory sorts into scaffolding vs.
test cases at a glance:

  harness_*_test.go / main_test.go  - support code (no test cases)
  filters_test.go, notifications_test.go - the actual TestXxx cases

Also rename the fake-notifier type recorder -> fakeNotifier, since "recorder"
didn't convey that it stands in for a notification server.

No behavioural change; same tests, regrouped.
